Mover serializaçao para gateway
